pyproject.toml
[tool.poetry]
name = "raskladochka"
version = "0.1.0"
description = ""
authors = ["andreyvaran <moscow.russia.mai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
langchain = "^0.1.14"
gigachat = "^0.1.22"
grpcio = "^1.62.1"
pydub = "^0.25.1"
ffmpeg = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
